+++ examples/miniapps/bundles/run.py
@@ -1,4 +1,7 @@
-"""Example application - Bundles."""
+"""Run 'Bundles' example application."""
+
+import sqlite3
+import boto3
 
 from dependency_injector import containers
 from dependency_injector import providers
@@ -10,15 +13,22 @@ from bundles.photos import Photos
 class Core(containers.DeclarativeContainer):
     """Core container."""
 
-    pgsql = providers.Singleton(object)
-    s3 = providers.Singleton(object)
+    config = providers.Configuration('config')
+    sqlite = providers.Singleton(sqlite3.connect, config.database.dsn)
+    s3 = providers.Singleton(
+        boto3.client, 's3',
+        aws_access_key_id=config.aws.access_key_id,
+        aws_secret_access_key=config.aws.secret_access_key)
 
 
 if __name__ == '__main__':
     # Initializing containers
     core = Core()
-    users = Users(database=core.pgsql)
-    photos = Photos(database=core.pgsql, file_storage=core.s3)
+    core.config.update({'database': {'dsn': ':memory:'},
+                        'aws': {'access_key_id': 'KEY',
+                                'secret_access_key': 'SECRET'}})
+    users = Users(database=core.sqlite)
+    photos = Photos(database=core.sqlite, file_storage=core.s3)
 
     # Fetching few users
     user_repository = users.user_repository()
@@ -28,4 +38,4 @@ if __name__ == '__main__':
     # Making some checks
     assert user1.id == 1
     assert user2.id == 2
-    assert user_repository.db is core.pgsql()
+    assert user_repository.db is core.sqlite()